requirements.txt
fastapi
uvicorn[standard]
python-chess
torch
numpy
